requirements.txt
streamlit
pytesseract
Pillow
ollama